feat: add AuthenticatableModel concern
